# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Naturaliza Api",
    "description": "this is api for a Ecommerse, that products are diferents types of mixed of teas",
    "contact": {
      "name": "Luis Lozano"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
    "/products/listproducts": {
      "get": {
        "tags": ["Products"],
        "summary": "Return all products not deleted",
        "description": "Present in Json format all products active in the database",
        "operationId": "listproducts",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Products"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/products/oneproduct/{code}": {
      "get": {
        "tags": ["Products"],
        "summary": "Find product by Code",
        "description": "For valid response try string Code",
        "operationId": "oneProduct",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Code of product that needs to be review",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "404": {
            "description": "Product not found with code: ???"
          },
          "500": {
            "description": "Error getProductByCode"
          }
        }
      }
    },
...
    "/products/createProduct": {
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product",
        "description": "Place a new product in the store",
        "operationId": "createNewProduct",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "characteristics": {
                    "type": "object",
                    "properties": {
                      "origin": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "flavor": {
                        "type": "string"
                      },
                      "properties": {
                        "type": "string"
                      },
                      "caffeineContent": {
                        "type": "string"
                      },
                      "allergens": {
                        "type": "string"
                      },
                      "organicCertification": {
                        "type": "boolean"
                      }
                    }
                  },
                  "ingredients": {
                    "type": "string"
                  },
                  "presentations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "size": {
                          "type": "string"
                        },
                        "price": {
                          "type": "number",
                          "format": "float"
                        },
                        "stock": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "preparationInstructions": {
                    "type": "object",
                    "properties": {
                      "recommendedPortion": {
                        "type": "string"
                      },
                      "waterTemperature": {
                        "type": "string"
                      },
                      "infusionTime": {
                        "type": "string"
                      },
                      "aditionalInfo": {
                        "type": "string"
                      }
                    }
                  }
                },
                "encoding": {
                  "image": {
                    "contentType": "image/*"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "400": {
            "description": "Errors in the validation of the data with zod"
          },
          "500": {
            "description": "Error createProduct:"
          }
        },
        "security": [
          {
            "user_auth_Sales": ["token Beaner"]
          }
        ]
      }
    },
...
    "/products/deleteProduct/{code}": {
      "delete": {
        "tags": ["Products"],
        "summary": "Delete product by Code",
        "description": "For valid response try code with value only string. nonstring will generate API errors",
        "operationId": "deleteProductByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Code of the product that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product Delete"
          },
          "404": {
            "description": "Product not found with code: {code}"
          },
          "500": {
            "description": "Error destroyProduct:"
          }
        },
        "security": [
          {
            "user_auth_Sales": ["token Beaner"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Products": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ADB145D"
          },
          "name": {
            "type": "string",
            "example": "Infusion de te verde"
          },
          "description": {
            "type": "string",
            "example": "Te traido de la india mezclado con variades del mismo"
          },
          "image": {
            "type": "string",
            "description": "url of the image in the bucket",
            "example": "https://xdcrhbhtzfatlmeoejjo.supabase.co/storage/v1/object/public/imagens/imagen/1739055989514_Rectangle_49.png"
          },
          "characteristics": {
            "type": "object",
            "properties": {
              "origin": {
                "type": "string",
                "example": "India"
              },
              "type": {
                "type": "string",
                "example": "Te verde"
              },
              "flavor": {
                "type": "string",
                "example": "Picante"
              },
              "properties": {
                "type": "string",
                "example": "Digestivo"
              },
              "caffeineContent": {
                "type": "string",
                "example": "Bajo contenido"
              },
              "allergens": {
                "type": "string",
                "example": "Sin Gluten"
              },
              "organicCertification": {
                "type": "boolean",
                "example": true
              }
            }
          },
          "ingredients": {
            "type": "string",
            "example": "Te verde, jengibre"
          },
          "presentations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "size": {
                  "type": "string",
                  "example": "Bolsa 50gr"
                },
                "price": {
                  "type": "number",
                  "format": "float",
                  "example": 15.23
                },
                "stock": {
                  "type": "integer",
                  "example": 125
                }
              }
            }
          },
          "preparationInstructions": {
            "type": "object",
            "properties": {
              "recommendedPortion": {
                "type": "string",
                "example": "30 gr"
              },
              "waterTemperature": {
                "type": "string",
                "example": "30 °C"
              },
              "infusionTime": {
                "type": "string",
                "example": "3 - 10 min"
              },
              "aditionalInfo": {
                "type": "string",
                "example": "Excelente Sabor"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "user_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "user_auth_Sales": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}